// src/Graphql/CodeGen/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Int or string */
  ActionCount: any;
  /** AreaOfEffectType ("SPHERE", "CUBE", "CYLINDER", "LINE", "CONE") or list of AreaOfEffectTypes */
  AreaOfEffectTypeFilter: any;
  /** Float, list of floats, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  FloatFilter: any;
  /** Int, list of ints, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  IntFilter: any;
  /** LanguageScript ("COMMON", "ELVISH", "DWARVISH", "INFERNAL", "DRACONIC", "CELESTIAL") or list of LanguageScripts */
  LanguageScriptFilter: any;
  /** MonsterSubtype ("ANY_RACE", "HUMAN", "DWARF", "ELF", "GOBLINOID", "MERFOLK", "SHAPECHANGER", "DEMON", "DEVIL", "ORC", "SAHUAGIN", "TITAN", "KOBOLD", "GNOLL", "GRIMLOCK", "LIZARDFOLK", "GNOME") or list of MonsterSubtypes */
  MonsterSubtypeFilter: any;
  /** MonsterTypes ("BEAST", "MONSTROSITY", "DRAGON", "HUMANOID", "UNDEAD", "FIEND", "CELESTIAL", "CONSTRUCT", "GIANT", "ELEMENTAL", "FEY", "ABERRATION", "OOZE", "SWARM", "PLANT") or list of MonsterTypes */
  MonsterTypeFilter: any;
  /** ProficiencyType ("WEAPONS", "ARTISANS_TOOLS", "SKILLS", "ARMOR", "MUSICAL_INSTRUMENTS", "SAVING_THROWS", "OTHER", "GAMING_SETS", "VEHICLES") or list of ProficiencyTypes */
  ProficiencyTypeFilter: any;
  /** Size ("TINY", "SMALL", "MEDIUM", "LARGE", "HUGE", "GARGANTUAN") or list of sizes */
  SizeFilter: any;
  /** SpellAttackType ("MELEE", "RANGED") or list of SpellAttackTypes */
  SpellAttackTypeFilter: any;
  /** String or list of strings */
  StringFilter: any;
};
// ...
export type Action = {
  __typename?: 'Action';
  action_name: Scalars['String'];
  count: Scalars['ActionCount'];
  type: Scalars['String'];
};

export type ActionDamage = {
  __typename?: 'ActionDamage';
  choose?: Maybe<Scalars['Int']>;
  damage_dice?: Maybe<Scalars['String']>;
  damage_type?: Maybe<DamageType>;
  dc?: Maybe<ActionDc>;
  from?: Maybe<DamageOptionSet>;
  type?: Maybe<Scalars['String']>;
};

export type ActionDc = {
  __typename?: 'ActionDc';
  success: DcSuccess;
  type: AbilityScore;
  value: Scalars['Int'];
};
// ...
export type Attack = {
  __typename?: 'Attack';
  damage?: Maybe<Array<Damage>>;
  dc: ActionDc;
  name: Scalars['String'];
};
// ...
export type BreathChoice = {
  __typename?: 'BreathChoice';
  choose: Scalars['Int'];
  from: BreathOptionSet;
  type: Scalars['String'];
};
// ...
export type Condition = {
  __typename?: 'Condition';
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
};

export type Cost = {
  __typename?: 'Cost';
  quantity: Scalars['Int'];
  unit: Currency;
};

export type CountedReferenceOption = {
  __typename?: 'CountedReferenceOption';
  count: Scalars['Int'];
  of: IEquipment;
  option_type: Scalars['String'];
  prerequisites?: Maybe<Array<ProficiencyPrerequisite>>;
};

export enum Currency {
  Cp = 'CP',
  Gp = 'GP',
  Sp = 'SP'
}

export type Damage = {
  __typename?: 'Damage';
  damage_dice: Scalars['String'];
  damage_type: DamageType;
};
// ...
export type LegendaryAction = {
  __typename?: 'LegendaryAction';
  damage?: Maybe<Array<Damage>>;
  dc?: Maybe<ActionDc>;
  desc: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type Monster = {
  __typename?: 'Monster';
  actions?: Maybe<Array<MonsterAction>>;
  armor_class: Scalars['Int'];
  challenge_rating: Scalars['Float'];
  charisma: Scalars['Int'];
  condition_immunities: Array<Condition>;
  constitution: Scalars['Int'];
  damage_immunities: Array<Scalars['String']>;
  damage_resistances: Array<Scalars['String']>;
  damage_vulnerabilities: Array<Scalars['String']>;
  desc?: Maybe<Scalars['String']>;
  dexterity: Scalars['Int'];
  forms?: Maybe<Array<Monster>>;
  hit_dice: Scalars['String'];
  hit_points: Scalars['Int'];
  hit_points_roll: Scalars['String'];
  image?: Maybe<Scalars['String']>;
  index: Scalars['String'];
  intelligence: Scalars['Int'];
  languages: Scalars['String'];
  legendary_actions?: Maybe<Array<LegendaryAction>>;
  name: Scalars['String'];
  proficiencies: Array<MonsterProficiency>;
  reactions?: Maybe<Array<Reaction>>;
  senses: Senses;
  size: Size;
  special_abilities?: Maybe<Array<SpecialAbility>>;
  speed: MonsterSpeed;
  strength: Scalars['Int'];
  subtype?: Maybe<MonsterSubtype>;
  type: MonsterType;
  wisdom: Scalars['Int'];
  xp: Scalars['Int'];
};

export type MonsterAction = {
  __typename?: 'MonsterAction';
  action_options?: Maybe<MonsterActionChoice>;
  actions?: Maybe<Array<Action>>;
  attack_bonus?: Maybe<Scalars['Int']>;
  attacks?: Maybe<Array<Attack>>;
  damage?: Maybe<Array<ActionDamage>>;
  dc?: Maybe<ActionDc>;
  desc: Scalars['String'];
  multiattack_type?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  options?: Maybe<BreathChoice>;
  usage?: Maybe<Usage>;
};

export type MonsterActionChoice = {
  __typename?: 'MonsterActionChoice';
  choose: Scalars['Int'];
  from: MonsterActionOptionSet;
  type: Scalars['String'];
};
// ...
export type MonsterProficiency = {
  __typename?: 'MonsterProficiency';
  proficiency: Proficiency;
  value: Scalars['Int'];
};

export type MonsterSpeed = {
  __typename?: 'MonsterSpeed';
  burrow?: Maybe<Scalars['String']>;
  climb?: Maybe<Scalars['String']>;
  fly?: Maybe<Scalars['String']>;
  hover?: Maybe<Scalars['Boolean']>;
  swim?: Maybe<Scalars['String']>;
  walk?: Maybe<Scalars['String']>;
};

export type MonsterSpell = {
  __typename?: 'MonsterSpell';
  spell: Spell;
  usage?: Maybe<Usage>;
};

export type MonsterSpellSlot = {
  __typename?: 'MonsterSpellSlot';
  level: Scalars['Int'];
  slots: Scalars['Int'];
};

export type MonsterSpellcasting = {
  __typename?: 'MonsterSpellcasting';
  ability: AbilityScore;
  components_required?: Maybe<Array<SpellComponent>>;
  dc?: Maybe<Scalars['Int']>;
  level?: Maybe<Scalars['Int']>;
  modifier?: Maybe<Scalars['Int']>;
  school?: Maybe<Scalars['String']>;
  slots?: Maybe<Array<MonsterSpellSlot>>;
  spells: Array<MonsterSpell>;
};

export enum MonsterSubtype {
  AnyRace = 'ANY_RACE',
  Demon = 'DEMON',
  Devil = 'DEVIL',
  Dwarf = 'DWARF',
  Elf = 'ELF',
  Gnoll = 'GNOLL',
  Gnome = 'GNOME',
  Goblinoid = 'GOBLINOID',
  Grimlock = 'GRIMLOCK',
  Human = 'HUMAN',
  Kobold = 'KOBOLD',
  Lizardfolk = 'LIZARDFOLK',
  Merfolk = 'MERFOLK',
  Orc = 'ORC',
  Sahuagin = 'SAHUAGIN',
  Shapechanger = 'SHAPECHANGER',
  Titan = 'TITAN'
}

export enum MonsterType {
  Aberration = 'ABERRATION',
  Beast = 'BEAST',
  Celestial = 'CELESTIAL',
  Construct = 'CONSTRUCT',
  Dragon = 'DRAGON',
  Elemental = 'ELEMENTAL',
  Fey = 'FEY',
  Fiend = 'FIEND',
  Giant = 'GIANT',
  Humanoid = 'HUMANOID',
  Monstrosity = 'MONSTROSITY',
  Ooze = 'OOZE',
  Plant = 'PLANT',
  Swarm = 'SWARM',
  Undead = 'UNDEAD'
}
// ...
export type Proficiency = {
  __typename?: 'Proficiency';
  classes: Array<Class>;
  index: Scalars['String'];
  name: Scalars['String'];
  races: Array<ProficiencyRace>;
  reference: ProficiencyReference;
  type: ProficiencyType;
};
// ...
export type Reaction = {
  __typename?: 'Reaction';
  dc?: Maybe<ActionDc>;
  desc: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type Senses = {
  __typename?: 'Senses';
  blindsight?: Maybe<Scalars['String']>;
  darkvision?: Maybe<Scalars['String']>;
  passive_perception: Scalars['Int'];
  tremorsense?: Maybe<Scalars['String']>;
  truesight?: Maybe<Scalars['String']>;
};

export enum Size {
  Gargantuan = 'GARGANTUAN',
  Huge = 'HUGE',
  Large = 'LARGE',
  Medium = 'MEDIUM',
  Small = 'SMALL',
  Tiny = 'TINY'
}
// ...
export type SpecialAbility = {
  __typename?: 'SpecialAbility';
  damage?: Maybe<Array<Damage>>;
  dc?: Maybe<ActionDc>;
  desc: Scalars['String'];
  name: Scalars['String'];
  spellcasting?: Maybe<MonsterSpellcasting>;
  usage?: Maybe<Usage>;
};
// ...
export type Usage = {
  __typename?: 'Usage';
  dice?: Maybe<Scalars['String']>;
  min_value?: Maybe<Scalars['Int']>;
  rest_types?: Maybe<Array<RestType>>;
  times?: Maybe<Scalars['Int']>;
  type: UsageType;
};